Settingsフォルダを監視するように修正
NSFileCoordinatorによる監視はまだ存在しないファイルが作成されることを監視するのが
できないようだったためSettingsフォルダを監視するように修正
